+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Development
 
+### Twin Peaks(1.0.2)
+- diffUtilCallback created
 ### Twin Peaks(1.0.1)
 - adaptersOf type corrected.
 ### Twin Peaks(1.0.0)